chore: Rename go.env to go.sh
This commit renames the file go.env to go.sh in the .config directory. The file defines the Go environment and is meant to be included in user or system files. The GOPATH variable is exported with the path to the Go directory. The changes were made to improve clarity and consistency in the codebase.

chore: Remove pnpm.env file

This commit removes the pnpm.env file from the .config directory. The file contained environment variables related to the PNPM package manager. Since the file was no longer necessary, it has been deleted.

feat: Add pnpm.sh file

This commit adds the pnpm.sh file to the .config directory. The file defines the PNPM_HOME variable and updates the PATH environment variable accordingly. This allows the PNPM package manager to be used from the command line. The changes were made to improve the functionality and usability of the code.

chore: Update .shellcheckrc file

This commit updates the .shellcheckrc file in the .trunk/configs directory. The changes disable specific ShellCheck rules (SC2148, SC2154, SC1090, SC1091) to improve code analysis. The source-path variable is set to SCRIPTDIR to specify the source path for the ShellCheck tool.

chore: Update trunk.yaml file

This commit updates the trunk.yaml file in the .trunk directory. The changes update the versions of various linters and code analysis tools (trivy@0.44.1, trufflehog@3.

Signed-off-by: Salvydas Lukosius <sall@w-ss.io>